Organize tx utils in shared
