+++ apps/desktop-wallet/src/features/transactionsDisplay/transactionDetailsModal/PendingSentAddressBadge.tsx
@@ -1,6 +1,4 @@
-import { AddressHash, selectPendingSentTransactionByHash } from '@alephium/shared'
-import { useTransactionDirection } from '@alephium/shared-react'
-import { explorer as e } from '@alephium/web3'
+import { selectPendingSentTransactionByHash } from '@alephium/shared'
 import { useTranslation } from 'react-i18next'
 
 import AddressBadge from '@/components/AddressBadge'
@@ -8,15 +6,12 @@ import Badge from '@/components/Badge'
 import { useAppSelector } from '@/hooks/redux'
 
 interface PendingSentAddressBadgeProps {
-  tx: e.Transaction | e.PendingTransaction
-  referenceAddress: AddressHash
-  isDestinationAddress?: boolean
+  txHash: string
 }
 
-const PendingSentAddressBadge = ({ tx, referenceAddress, isDestinationAddress }: PendingSentAddressBadgeProps) => {
+const PendingSentAddressBadge = ({ txHash }: PendingSentAddressBadgeProps) => {
   const { t } = useTranslation()
-  const pendingSentTx = useAppSelector((s) => selectPendingSentTransactionByHash(s, tx.hash))
-  const direction = useTransactionDirection(tx, referenceAddress)
+  const pendingSentTx = useAppSelector((s) => selectPendingSentTransactionByHash(s, txHash))
 
   if (!pendingSentTx) return null
 
@@ -24,12 +19,7 @@ const PendingSentAddressBadge = ({ tx, referenceAddress, isDestinationAddress }:
 
   if (pendingSentTx.type === 'faucet') return <Badge>{t('Token faucet')}</Badge>
 
-  return (
-    <AddressBadge
-      truncate
-      addressHash={isDestinationAddress || direction !== 'in' ? pendingSentTx.toAddress : pendingSentTx.fromAddress}
-    />
-  )
+  return <AddressBadge truncate addressHash={pendingSentTx.toAddress} />
 }
 
 export default PendingSentAddressBadge
