--- packages/shared/src/transactions.ts
@@ -1,260 +1,0 @@
-import { ALPH } from '@alephium/token-list'
-import {
-  explorer as e,
-  isContractAddress,
-  isGrouplessAddress,
-  isGrouplessAddressWithGroupIndex,
-  isGrouplessAddressWithoutGroupIndex,
-  node as n
-} from '@alephium/web3'
-
-import { AddressHash } from '@/types/addresses'
-import { AssetAmount } from '@/types/assets'
-import { SignTransferTxModalProps } from '@/types/signTxModalTypes'
-import { AmountDeltas, SentTransaction, TransactionDirection } from '@/types/transactions'
-import { uniq } from '@/utils'
-
-// Address without group number
-export const getBaseAddressStr = (address: string): string => {
-  const parts = address.split(':')
-  return parts[0]
-}
-
-export const getInputOutputBaseAddress = (io: e.Input | e.Output): string | undefined =>
-  io.address ? getBaseAddressStr(io.address) : undefined
-
-export const isSameBaseAddress = (address1: string, address2: string): boolean =>
-  getBaseAddressStr(address1) === getBaseAddressStr(address2)
-
-export const calcTxAmountsDeltaForAddress = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction,
-  refAddress: string
-): AmountDeltas => {
-  if (!tx.inputs || !tx.outputs) throw 'Missing transaction details'
-
-  if (getTxAddresses(tx).every((address) => getBaseAddressStr(address) === refAddress)) {
-    const totalInputAlph = tx.inputs.reduce((sum, i) => sum + BigInt(i.attoAlphAmount ?? 0), BigInt(0))
-    const totalOutputAlph = tx.outputs.reduce((sum, o) => sum + BigInt(o.attoAlphAmount ?? 0), BigInt(0))
-    const fee = totalOutputAlph - totalInputAlph
-
-    return {
-      alphAmount: fee,
-      tokenAmounts: []
-    }
-  }
-
-  const outputAmounts = sumUpAddressInputOutputAmounts(refAddress, tx.outputs)
-  const inputAmounts = sumUpAddressInputOutputAmounts(refAddress, tx.inputs)
-
-  const tokensDelta = [...outputAmounts.tokenAmounts]
-
-  inputAmounts.tokenAmounts.forEach((inputToken) => {
-    const existingTokenDelta = tokensDelta.find(({ id }) => id === inputToken.id)
-    if (existingTokenDelta) {
-      existingTokenDelta.amount -= inputToken.amount
-    } else {
-      tokensDelta.push({
-        id: inputToken.id,
-        amount: -inputToken.amount
-      })
-    }
-  })
-
-  return {
-    alphAmount: outputAmounts.alphAmount - inputAmounts.alphAmount,
-    tokenAmounts: tokensDelta.filter(({ amount }) => amount !== BigInt(0))
-  }
-}
-
-const sumUpAddressInputOutputAmounts = (refAddress: string, io: (e.Input | e.Output)[]) => {
-  const isGrouplessRefAddress = isGrouplessAddressWithoutGroupIndex(refAddress)
-
-  return io.reduce(
-    (acc, io) => {
-      if (!io.address) return acc
-
-      const isNotSameAddress = isGrouplessRefAddress
-        ? !isSameBaseAddress(io.address, refAddress)
-        : io.address !== refAddress
-
-      if (isNotSameAddress) return acc
-
-      acc.alphAmount += BigInt(io.attoAlphAmount ?? 0)
-
-      if (!io.tokens) return acc
-
-      io.tokens.forEach((token) => {
-        const existingToken = acc.tokenAmounts.find((t) => t.id === token.id)
-        existingToken
-          ? (existingToken.amount += BigInt(token.amount))
-          : acc.tokenAmounts.push({
-              id: token.id,
-              amount: BigInt(token.amount)
-            })
-      })
-
-      return acc
-    },
-    { alphAmount: BigInt(0), tokenAmounts: [] } as AmountDeltas
-  )
-}
-
-// TODO: Clean up use of Transaction | PendingTransaction | MempoolTransaction
-
-export const getDirection = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction,
-  address: string
-): TransactionDirection => (calcTxAmountsDeltaForAddress(tx, address).alphAmount < 0 ? 'out' : 'in')
-
-export const isConsolidationTx = (tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction): boolean => {
-  const inputAddresses = tx.inputs ? uniq(tx.inputs.map((input) => input.address)) : []
-  const outputAddresses = tx.outputs ? uniq(tx.outputs.map((output) => output.address)) : []
-
-  return (
-    inputAddresses.length === 1 &&
-    outputAddresses.length === 1 &&
-    inputAddresses[0] !== undefined &&
-    outputAddresses[0] !== undefined &&
-    isSameBaseAddress(inputAddresses[0], outputAddresses[0])
-  )
-}
-
-export const isSelfTransfer = (tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction): boolean => {
-  const inputAddresses = tx.inputs ? uniq(tx.inputs.map((input) => input.address)) : []
-  const outputAddresses = tx.outputs ? uniq(tx.outputs.map((output) => output.address)) : []
-
-  return (
-    inputAddresses.length === 1 &&
-    outputAddresses.length === 1 &&
-    inputAddresses[0] !== undefined &&
-    outputAddresses[0] !== undefined &&
-    inputAddresses[0] === outputAddresses[0]
-  )
-}
-
-export const isAlphAmountReduced = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction,
-  refAddress: string
-): boolean => {
-  const { alphAmount } = calcTxAmountsDeltaForAddress(tx, refAddress)
-
-  return alphAmount < 0
-}
-
-export const isConfirmedTx = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction | SentTransaction
-): tx is e.Transaction =>
-  // See https://github.com/alephium/alephium-frontend/issues/1367
-  'blockHash' in tx && !tx.inputs?.some((input) => input.txHashRef === undefined)
-
-export const isRichTransaction = (
-  tx: e.PendingTransaction | n.RichTransaction | e.AcceptedTransaction | e.Transaction
-): tx is n.RichTransaction => 'unsigned' in tx
-
-export const isSentTx = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction | SentTransaction
-): tx is SentTransaction => 'status' in tx
-
-export const isWalletSelfTransfer = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction,
-  walletAddresses: AddressHash[]
-): boolean => getTxAddresses(tx).every((address) => walletAddresses.includes(getBaseAddressStr(address)))
-
-export const isBidirectionalTransfer = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction,
-  referenceAddress: string
-): boolean => {
-  const { alphAmount, tokenAmounts } = calcTxAmountsDeltaForAddress(tx, referenceAddress)
-
-  return hasPositiveAndNegativeAmounts(alphAmount, tokenAmounts)
-}
-
-export const isContractTx = (tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction): boolean =>
-  !!tx.outputs?.some(inputOutputIsContractAddress) || !!tx.inputs?.some(inputOutputIsContractAddress)
-
-const inputOutputIsContractAddress = (io: e.Input | e.Output): boolean => !!io.address && isContractAddress(io.address)
-
-export const getTxAddresses = (tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction): AddressHash[] => {
-  const addresses = new Set<AddressHash>()
-
-  for (const { address } of [...(tx.outputs ?? []), ...(tx.inputs ?? [])]) {
-    if (address) addresses.add(address)
-  }
-
-  return Array.from(addresses)
-}
-
-export const isGrouplessAddressIntraTransfer = (
-  tx: e.Transaction | e.PendingTransaction | e.MempoolTransaction
-): boolean => {
-  const txAddresses = getTxAddresses(tx)
-  const firstAddress = txAddresses.at(0)
-
-  if (!firstAddress || !isGrouplessAddress(firstAddress)) return false
-
-  const firstBaseAddress = getBaseAddressStr(firstAddress)
-
-  return txAddresses.every(
-    (address) => isGrouplessAddressWithGroupIndex(address) && getBaseAddressStr(address) === firstBaseAddress
-  )
-}
-
-export const hasPositiveAndNegativeAmounts = (alphAmout: bigint, tokensAmount: Required<AssetAmount>[]): boolean => {
-  const allAmounts = [alphAmout, ...tokensAmount.map((tokenAmount) => tokenAmount.amount)]
-  const allAmountsArePositive = allAmounts.every((amount) => amount >= 0)
-  const allAmountsAreNegative = allAmounts.every((amount) => amount <= 0)
-
-  return !allAmountsArePositive && !allAmountsAreNegative
-}
-
-export const findTransactionReferenceAddress = (addresses: AddressHash[], tx: e.Transaction | e.PendingTransaction) =>
-  addresses.find((address) => isAddressPresentInInputsOutputs(address, tx))
-
-export const isAddressPresentInInputsOutputs = (
-  addressHash: AddressHash,
-  tx: e.Transaction | e.PendingTransaction | n.RichTransaction
-) => {
-  const inputs = isRichTransaction(tx) ? tx.unsigned.inputs : tx.inputs
-  const outputs = isRichTransaction(tx) ? tx.unsigned.fixedOutputs : tx.outputs
-
-  return (
-    inputs?.some((input) => input.address && isSameBaseAddress(input.address, addressHash)) ||
-    outputs?.some((output) => output.address && isSameBaseAddress(output.address, addressHash))
-  )
-}
-
-export const findTransactionInternalAddresses = (addresses: AddressHash[], tx: e.Transaction | n.RichTransaction) =>
-  addresses.filter((addressHash) => isAddressPresentInInputsOutputs(addressHash, tx))
-
-export const calculateTransferTxAssetAmounts = (txParams: SignTransferTxModalProps['txParams']) => {
-  const assetAmounts = [] as Required<AssetAmount>[]
-
-  const res = txParams.destinations.reduce(
-    (acc, destination) => {
-      acc.attoAlphAmount += BigInt(destination.attoAlphAmount)
-
-      destination.tokens?.forEach((token) => {
-        const t = acc.tokens.find(({ id }) => id === token.id)
-
-        if (t) {
-          t.amount += BigInt(token.amount)
-        } else {
-          acc.tokens.push({ id: token.id, amount: BigInt(token.amount) })
-        }
-      })
-
-      return acc
-    },
-    { attoAlphAmount: BigInt(0), tokens: [] as { id: string; amount: bigint }[] }
-  )
-
-  if (res.attoAlphAmount > 0) {
-    assetAmounts.push({ id: ALPH.id, amount: res.attoAlphAmount })
-  }
-
-  if (res.tokens.length > 0) {
-    assetAmounts.push(...res.tokens)
-  }
-
-  return assetAmounts
-}
