+++ packages/shared/src/utils/index.ts
@@ -2,6 +2,5 @@ export * from '@/utils/addresses'
 export * from '@/utils/utils'
 export * from '@/utils/walletConnect'
 export * from '@/utils/types'
-export * from '@/utils/transactions'
 export * from '@/utils/network'
 export * from '@/utils/errors'
